Restructuring accordion for Compensation and Search
